+++ R/method_mlpack.R
@@ -12,6 +12,7 @@
 #' @param k Number of neighbors to return.
 #' @param verbose If TRUE, log messages to the console.
 #' @param seed seed for the pseudo-random numbers algorithm.
+#' @param path path to write the index.
 #' @param control controls for  \code{lsh} or \code{kd}.
 #'
 #' @description
@@ -23,6 +24,7 @@ method_mlpack <- function(x,
                           k,
                           verbose,
                           seed,
+                          path,
                           control) {
 
   ## TODO: separate building an index from querying
@@ -58,9 +60,25 @@ method_mlpack <- function(x,
                                        rho = control$kd$rho,
                                        tau = control$kd$tau,
                                        random_basis = control$kd$random_basis))
+  if (!is.null(path)) {
+    if (grepl("(/|\\\\)$", path)) {
+      path_ann <- paste0(path, "index.annoy")
+      path_ann_cols <- paste0(path, "index-colnames.txt")
+    } else {
+      path_ann <- paste0(path, "//index.annoy")
+      path_ann_cols <- paste0(path, "//index-colnames.txt")
+    }
+    if (verbose == 2) {
+      cat("Writing an index to `path`\n")
+    }
+    l_ind$save(path_ann)
+    writeLines(colnames(x), path_ann_cols)
+  }
+
 
   l_df <- data.table::data.table(y = 1:NROW(y),
-                                 x = result$neighbors[, k] + 1)
+                                 x = result$neighbors[, k] + 1,
+                                 dist = result$distances[, k])
   l_df
 }