+++ inst/tinytest/test_annoy.R
@@ -17,7 +17,8 @@ expect_equal(
   structure(list(result = structure(
     list(x = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
          y = c(5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L),
-         block = c(2, 2, 2, 2, 1, 1, 1, 1)),
+         block = c(2, 2, 2, 2, 1, 1, 1, 1),
+         dist = c(0, 1, 0, 6, 1, 0, 1, 4)),
     row.names = c(NA, -8L),
     class = c("data.table", "data.frame")),
     method = "annoy",
@@ -41,7 +42,8 @@ expect_equal(
   structure(list(result = structure(
     list(x = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
          y = c(5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L),
-         block = c(2, 2, 2, 2, 1, 1, 1, 1)),
+         block = c(2, 2, 2, 2, 1, 1, 1, 1),
+         dist = c(0, 1, 0, 6, 1, 0, 1, 4)),
     row.names = c(NA, -8L),
     class = c("data.table", "data.frame")),
     method = "annoy",
@@ -59,3 +61,24 @@ expect_equal(
            distance = "euclidean")$result$block,
   c(1, 1, 1, 1, 2, 2, 2, 2)
 )
+
+
+## file saving
+
+expect_true({
+  blocking(x = mat_y,
+           ann = "annoy",
+           distance = "euclidean",
+           ann_write = ".")
+  file.exists("./index.annoy") &
+    file.exists("./index-colnames.txt")
+})
+
+## testing reading saved index
+expect_equal({
+  ncols <- length(readLines("./index-colnames.txt"))
+  ann_annoy <- methods::new(RcppAnnoy::AnnoyManhattan, ncols)
+  ann_annoy$load("./index.annoy")
+  ann_annoy$getNItems()
+},  8)
+