+++ inst/tinytest/test_mlpack.R
@@ -7,7 +7,8 @@ expect_equal(
   structure(list(result = structure(
     list(x = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
          y = c(5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L),
-         block = c(2, 2, 2, 2, 1, 1, 1, 1)),
+         block = c(2, 2, 2, 2, 1, 1, 1, 1),
+         dist = c(0, 1, 0, 2.44948974278318, 1, 0, 1, 2)),
     row.names = c(NA, -8L),
     class = c("data.table", "data.frame")),
     method = "lsh",
@@ -25,7 +26,8 @@ expect_equal(
   structure(list(result = structure(
     list(x = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
          y = c(5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L),
-         block = c(2, 2, 2, 2, 1, 1, 1, 1)),
+         block = c(2, 2, 2, 2, 1, 1, 1, 1),
+         dist = c(0, 1, 0, 2.44948974278318, 1, 0, 1, 2)),
     row.names = c(NA, -8L),
     class = c("data.table", "data.frame")),
     method = "kd",
@@ -56,7 +58,8 @@ expect_equal(
   structure(list(result = structure(
     list(x = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
          y = c(5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L),
-         block = c(2, 2, 2, 2, 1, 1, 1, 1)),
+         block = c(2, 2, 2, 2, 1, 1, 1, 1),
+         dist = c(0, 1, 0, 2.44948974278318, 1, 0, 1, 2)),
     row.names = c(NA, -8L),
     class = c("data.table", "data.frame")),
     method = "lsh",
@@ -75,7 +78,8 @@ expect_equal(
   structure(list(result = structure(
     list(x = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
          y = c(5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L),
-         block = c(2, 2, 2, 2, 1, 1, 1, 1)),
+         block = c(2, 2, 2, 2, 1, 1, 1, 1),
+         dist = c(0, 1, 0, 2.44948974278318, 1, 0, 1, 2)),
     row.names = c(NA, -8L),
     class = c("data.table", "data.frame")),
     method = "kd",