+++ NEWS.md
@@ -9,3 +9,4 @@
 7. s3method for printing.
 8. first vignette added.
 9. evaluation with standard metrics (recall, fpr etc) added, works with vector for deduplication.
+10. added saving index for hnsw and annoy